refactor(notes,files): extract note_type, notes, file_uploads, reports into dojo/<module>/ [Phase 1,3,5,6,8,9]
