+++ dojo/file_uploads/__init__.py
@@ -1,0 +1,1 @@
+import dojo.file_uploads.admin  # noqa: F401
