+++ dojo/note_type/__init__.py
@@ -1,0 +1,1 @@
+import dojo.note_type.admin  # noqa: F401
