+++ dojo/note_type/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.note_type.models import Note_Type
+
+admin.site.register(Note_Type)
