+++ dojo/note_type/api/__init__.py
@@ -1,0 +1,1 @@
+path = "note_type"  # noqa: RUF067
