+++ dojo/notes/__init__.py
@@ -1,0 +1,1 @@
+import dojo.notes.admin  # noqa: F401
