+++ dojo/notes/api/__init__.py
@@ -1,0 +1,1 @@
+path = "notes"  # noqa: RUF067
