+++ dojo/reports/__init__.py
@@ -1,0 +1,1 @@
+import dojo.reports.admin  # noqa: F401
