+++ dojo/reports/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.reports.models import Report_Type
+
+admin.site.register(Report_Type)
