+++ dojo/reports/models.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+
+class Report_Type(models.Model):
+    name = models.CharField(max_length=255)
