+++ dojo/reports/ui/views.py
@@ -27,13 +27,13 @@ from dojo.finding.ui.filters import (
     ReportFindingFilterWithoutObjectLookups,
 )
 from dojo.finding.ui.views import BaseListFindings
-from dojo.forms import ReportOptionsForm
 from dojo.labels import get_labels
 from dojo.location.models import Location
 from dojo.location.queries import get_authorized_locations
 from dojo.location.status import FindingLocationStatus
 from dojo.models import Dojo_User, Endpoint, Engagement, Finding, Product, Product_Type, Test
 from dojo.reports.queries import prefetch_related_endpoints_for_report, prefetch_related_findings_for_report
+from dojo.reports.ui.forms import ReportOptionsForm
 from dojo.reports.widgets import (
     CoverPage,
     CustomReportJsonForm,
